Refactor location components with improved skeleton loading and Dutch translations
- Extract location page logic into dedicated feature pages
- Update LocationSkeleton with cleaner design matching FeaturedLocationsSection
- Refactor FeaturedLocationsSection to use LocationGrid component
- Make skeleton loader count configurable via skeletonCount prop
- Translate Dutch location titles and subtitles in nl.json
- Improve component reusability and maintainability
