+++ frontend/src/features/locations/components/LocationSkeleton.tsx
@@ -1,40 +1,12 @@
-import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card';
-import { Skeleton } from '@/components/ui/skeleton';
-
 export function LocationSkeleton() {
   return (
-    <Card>
-      <CardHeader>
-        <div className="flex items-center gap-2">
-          <Skeleton className="h-5 w-5" />
-          <Skeleton className="h-6 w-3/4" />
-        </div>
-        <Skeleton className="h-4 w-full" />
-      </CardHeader>
-      <CardContent className="space-y-4">
-        <div className="space-y-2">
-          <Skeleton className="h-4 w-full" />
-          <Skeleton className="h-4 w-2/3" />
-        </div>
-        
-        <div className="flex items-center justify-between">
-          <div className="flex items-center gap-2">
-            <Skeleton className="h-4 w-4" />
-            <Skeleton className="h-4 w-20" />
-          </div>
-          <Skeleton className="h-5 w-16" />
-        </div>
-        
-        <div className="flex items-center gap-2">
-          <Skeleton className="h-4 w-4" />
-          <Skeleton className="h-4 w-24" />
-        </div>
-        
-        <div className="flex gap-2 pt-2">
-          <Skeleton className="h-8 flex-1" />
-          <Skeleton className="h-8 flex-1" />
-        </div>
-      </CardContent>
-    </Card>
+    <div className="w-full bg-white rounded-lg shadow-lg overflow-hidden">
+      <div className="h-48 bg-gray-200 animate-pulse" />
+      <div className="p-6 space-y-4">
+        <div className="h-6 bg-gray-200 animate-pulse rounded" />
+        <div className="h-4 bg-gray-200 animate-pulse rounded w-2/3" />
+        <div className="h-4 bg-gray-200 animate-pulse rounded w-1/2" />
+      </div>
+    </div>
   );
 }
