+++ frontend/src/i18n/messages/nl.json
@@ -77,8 +77,8 @@
     "logo": "Boop-a-Booking"
   },
   "locations": {
-    "title": "Stay at our top unique properties",
-    "subtitle": "From castles and villas to boats and igloos, we have it all"
+    "title": "Verblijf in onze meest unieke accommodaties",
+    "subtitle": "Van kastelen en villa's tot boten en iglo's, wij hebben het allemaal"
   },
   "featuredLocations": {
     "title": "Uitgelichte Bestemmingen",
